Add toggle, reposition on show to TFWidget; fix DestroyWindow().
* Add TFWidget::ToggleShow() to swap between hidden and visible widget state.
* Add TFWidget::AuditWindowGeometry() to force root window centered for VR or when window
is initially created. For all other cases, if screen geometry would make window invisible, force it
onto the screen enough that the user can clearly find it.
* Modify TFWidget::Show() to call AuditWindowGeometry() when showing a root widget/window.
* Modify MenuHandler() in LiveTraffic.cpp to replace custom toggle logic by calling settingsUI.ToggleShow().
* Modify DestroyWindow() in TextIO.cpp to correctly set g_window to 0.

I'm sure I missed something, but this covers most of what I was thinking to address the re-centering vs. toggling the window issue.