Changed camera dataRefs to shared dataRefs, closes #196 again
Shared dataRefs allow the camera plugin to register a notification callback. It allows other plugins to use the same technique to inform a camera plugin. And it allows the camera plugin to reset the info to zero to indicate when a user stopped a camera view and returned to something else.